Fixed memory management, refactored print/bounds funcs, updated docs

+++ H2MM_C/rho_calc.c
@@ -4,27 +4,28 @@
 // Date created: 13 Feb 2021
 // Date modified: 14 Oct 2022
 
+#include <stdio.h>
+#include <stdlib.h>
+#include <stdint.h>
+#include <math.h>
+#include <time.h>
+
 #if defined(__linux__) || defined(__APPLE__)
-#include <unistd.h>
-//#include <pthread.h>
+#include <pthread.h>
 #elif _WIN32
 #include <windows.h>
 #endif
 
-#include <stdio.h>
-#include <stdlib.h>
-#include <math.h>
-#include <time.h>
 #include "C_H2MM.h"
 
 #define TRUE 1
 #define FALSE 0
 
 // calculates just the powers of the trans matrix up to maxdif
-trpow* transpow(unsigned long nstate, unsigned long maxdif, double* trans)
+trpow* transpow(int64_t nstate, int32_t maxdif, double* trans)
 {
 	// initialize variables
-	unsigned long i, j, k, t, istride, tstride, tstride_r, tistride, Ad;
+	int64_t i, j, k, t, istride, tstride, tstride_r, tistride, Ad;
 	double runsum;
 	trpow *power = (trpow*) malloc(sizeof(trpow));
 	
@@ -70,15 +71,15 @@ void* rhoulate(void *vals)
 {
 	// initialize variables
 	pwrs *D = (pwrs*) vals;
-	unsigned long i, j, k, m, z;
-	unsigned long Ad, Av, Rhod, Rhov;
-	unsigned long sdr = D->sT * D->td;
-	unsigned long sda = D->sj * D->td;
-	unsigned long svr = D->sT * D->tv;
-	unsigned long sva = D->sj * D->tv;
-	unsigned long sqr = D->sT * D->tq;
-	unsigned long sqa = D->sj * D->tq;
-	unsigned long si, sj, sk;
+	int64_t i, j, k, m, z;
+	int64_t Ad, Av, Rhod, Rhov;
+	int64_t sdr = D->sT * D->td;
+	int64_t sda = D->sj * D->td;
+	int64_t svr = D->sT * D->tv;
+	int64_t sva = D->sj * D->tv;
+	int64_t sqr = D->sT * D->tq;
+	int64_t sqa = D->sj * D->tq;
+	int64_t si, sj, sk;
 	double normsum;
 	// calculate power of trans matrix
 	for ( k = 0; k < D->sk; k++)
@@ -137,9 +138,9 @@ void* rhoulate(void *vals)
 	//pthread_exit(0);
 }
 
-void* rho_all(unsigned long nstate, double* transmat, pwrs *powers)
+void* rho_all(int64_t nstate, double* transmat, pwrs *powers)
 {
-	unsigned long i, j;
+	int64_t i, j;
 	// free A and Rho just in case
 	for ( i = 0; i < powers->max_pow * powers->sj; i++) powers->A[i] = 0.0;
 	for ( i = 0; i < powers->max_pow * powers->sT; i++) powers->Rho[i] = 0.0;
